vispy: make VispyWidget a first class object
